requirements.txt
discord.py~=2.4.0
pymongo~=4.10.1
python-dotenv~=1.0.1
requests
sentry-sdk~=2.21.0
urllib3~=2.3.0
decouple
python-dotenv
jishaku
fastapi
uvicorn
infisicalsdk
roblox~=2.0.0
openai~=1.66.3
bloxlink.py
cloudflare
discord~=2.3.2